+++ hosts/echo/darwin-configuration.nix
@@ -14,4 +14,9 @@
     publicHostKey =
       "c3NoLWVkMjU1MTkgQUFBQUMzTnphQzFsWkRJMU5URTVBQUFBSUpCV2N4Yi9CbGFxdDFhdU90RStGOFFVV3JVb3RpQzVxQkorVXVFV2RWQ2Igcm9vdEBuaXhvcwo=";
   }];
+
+  system.defaults.dock.persistent-apps = [
+    "/System/Volumes/Preboot/Cryptexes/App/System/Applications/Safari.app"
+    "/System/Applications/System Settings.app"
+  ];
 }